update from master branch

+++ LevelComponents/Room.h
@@ -155,7 +155,7 @@ namespace LevelComponents
         void SetBGLayerAutoScrollEnabled(bool enability);
         int GetLayerEffectsParam() { return (int) RoomHeader.LayerEffects; }
         LevelComponents::Door *GetDoor(int _doorID) { return doors[_doorID]; }
-        int CountDoors() { return doors.size(); }
+        int CountDoors() { return (int)doors.size(); }
         void SetDoors(std::vector<Door*> _doors) { doors = _doors; }
         int GetLocalDoorID(int globalDoorId);
         int GetCurrentEntitySetID() { return CurrentEntitySetID; }
